pyproject.toml
[tool.poetry]
name = "codequest_stats"
version = "0.1.0"
description = ""
authors = ["Savith Panamgama <geethikasavith@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.44.1"
pandas = "^2.2.3"
matplotlib = "^3.10.1"
google-api-python-client = "^2.167.0"
google-auth = "^2.39.0"
google-auth-oauthlib = "^1.2.1"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
